requirements.txt
python-telegram-bot==20.7
google-generativeai>=0.3.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
redis>=5.0.0
aiofiles>=23.0.0
python-dateutil>=2.8.0
pytz>=2023.3
cryptography>=41.0.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

